# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "커뮤니티 인기글 모아보기 API",
    "description": "여러 온라인 커뮤니티의 인기글을 모아서 한 곳에서 볼 수 있는 PostgREST 기반 API",
    "version": "2.0.0",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "로컬 개발 서버"
    },
    {
      "url": "https://api.example.com",
      "description": "프로덕션 서버"
    }
  ],
...
  "paths": {
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "게시글 목록 조회",
        "description": "## 여러 커뮤니티의 인기글을 조회합니다\n\n필터링, 정렬, 페이지네이션을 지원하는 포괄적인 게시글 조회 엔드포인트입니다.\n\n### 사용 사례\n- 🏠 홈페이지의 게시글 피드 로드\n- 🏷️ 특정 카테고리별 필터링 (기술, 뉴스 등)\n- 👍 추천순/최신순 정렬로 인기글 강조\n- 🔍 제목 및 내용 검색\n- 📄 페이지네이션으로 무한 스크롤 구현\n\n### 주의사항\n- **limit 최대값**: 100입니다\n- **Prefer 헤더**: `count=exact`를 설정하면 전체 개수를 `Content-Range` 헤더에서 받을 수 있습니다\n- **관계 데이터**: `select` 파라미터로 관계된 카테고리, 출처 정보를 함께 조회할 수 있습니다",
        "externalDocs": {
          "description": "고급 쿼리 예제 및 자세한 사용법 보기",
          "url": "https://github.com/yourrepo/docs/API_GUIDE.md#posts"
        },
        "operationId": "getPosts",
        "x-code-samples": [
          {
            "lang": "cURL",
            "summary": "기본 조회 (최신순, 6개)",
            "source": "curl -X GET 'http://localhost:3001/posts?limit=6&order=created_at.desc' \\\n  -H 'Prefer: count=exact'"
          },
          {
            "lang": "cURL",
            "summary": "기술 카테고리 필터 + 추천순 정렬",
            "source": "curl -X GET 'http://localhost:3001/posts?limit=10&categories.name=eq.tech&order=likes.desc'"
          },
          {
            "lang": "cURL",
            "summary": "AI 관련 검색",
            "source": "curl -X GET 'http://localhost:3001/posts?or=(title.ilike.*AI*,excerpt.ilike.*AI*)&limit=20'"
          },
          {
            "lang": "cURL",
            "summary": "페이지네이션 (2페이지, 6개씩)",
            "source": "curl -X GET 'http://localhost:3001/posts?limit=6&offset=6&order=created_at.desc'"
          },
          {
            "lang": "JavaScript",
            "summary": "Fetch API로 최신 게시글 로드",
            "source": "async function fetchLatestPosts(page = 1) {\n  const limit = 6;\n  const offset = (page - 1) * limit;\n  const response = await fetch(\n    `http://localhost:3001/posts?limit=${limit}&offset=${offset}&order=created_at.desc`,\n    { headers: { 'Prefer': 'count=exact' } }\n  );\n  const data = await response.json();\n  const total = response.headers.get('content-range')?.split('/')[1] || 0;\n  return { posts: data, totalCount: parseInt(total) };\n}"
          },
          {
            "lang": "JavaScript",
            "summary": "카테고리별 필터링",
            "source": "async function fetchByCategory(categoryName, limit = 10) {\n  const response = await fetch(\n    `http://localhost:3001/posts?categories.name=eq.${categoryName}&limit=${limit}&order=likes.desc`\n  );\n  return response.json();\n}"
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "한 페이지당 조회할 게시글 수\n\n- 기본값: 6\n- 최소값: 1\n- 최대값: 100\n- 팁: 무한 스크롤은 보통 10-20으로 설정",
            "example": 6,
            "schema": {
              "type": "integer",
              "default": 6,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "건너뛸 게시글 수 (페이지네이션용)\n\n- 예: offset=0 은 첫 번째부터\n- 예: offset=6 은 7번째부터 시작\n- 페이지 계산: `offset = (pageNumber - 1) * limit`",
            "example": 0,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "정렬 기준 (기본값: created_at.desc 최신순)\n\n**최신순 정렬**\n- `created_at.desc` - 최신글 먼저\n- `created_at.asc` - 오래된글 먼저\n\n**인기순 정렬**\n- `likes.desc` - 추천수 높은순\n- `likes.asc` - 추천수 낮은순\n\n**댓글순 정렬**\n- `comments.desc` - 댓글 많은순\n- `comments.asc` - 댓글 적은순",
            "example": "created_at.desc",
            "schema": {
              "type": "string",
              "enum": [
                "created_at.desc",
                "created_at.asc",
                "likes.desc",
                "likes.asc",
                "comments.desc",
                "comments.asc"
              ],
              "default": "created_at.desc"
            }
          },
          {
            "name": "categories.name",
            "in": "query",
            "description": "카테고리별 필터링\n\n`eq.` 연산자로 정확히 일치하는 카테고리만 조회\n\n**가능한 카테고리**\n- `tech` - 기술, 프로그래밍, AI, 클라우드\n- `news` - 뉴스, 정치, 경제\n- `entertainment` - 엔터테인먼트, 영화, 음악\n- `sports` - 스포츠, 게임\n- `life` - 생활, 건강, 음식\n\n**사용 예**\n- 기술만: `categories.name=eq.tech`\n- 뉴스만: `categories.name=eq.news`",
            "example": "tech",
            "schema": {
              "type": "string",
              "enum": ["tech", "news", "entertainment", "sports", "life"]
            }
          },
          {
            "name": "or",
            "in": "query",
            "description": "제목 또는 요약에서 검색 (대소문자 무시)\n\n**구문**: `or=(필드.연산자.*검색어*,필드.연산자.*검색어*)`\n\n**주요 연산자**\n- `ilike.*검색어*` - 포함 여부 검색\n\n**사용 예**\n- AI 검색: `or=(title.ilike.*AI*,excerpt.ilike.*AI*)`\n- React 관련: `or=(title.ilike.*React*,excerpt.ilike.*React*)`\n- 여러 키워드: `or=(title.ilike.*AI*,title.ilike.*ML*)`",
            "example": "(title.ilike.*AI*,excerpt.ilike.*AI*)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "in": "query",
            "description": "반환할 컬럼 선택 (기본: 모든 컬럼 + 관계 데이터)\n\n**기본 동작**: 생략하면 id,title,excerpt,likes,comments,created_at,url과 관계된 categories, sources 정보 자동 포함\n\n**사용 사례**\n- 특정 컬럼만: `id,title,likes`\n- 관계 데이터 포함: `id,title,categories(name,icon),sources(name)`\n\n**팁**: 필요한 컬럼만 선택하면 응답 크기를 줄일 수 있습니다",
            "example": "id,title,excerpt,likes,comments,created_at,url,categories(name,icon),sources(name,icon_emoji)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Prefer",
            "in": "header",
            "description": "응답 옵션 (정수 count 포함 여부)\n\n**count=exact**\n- Content-Range 헤더에 전체 개수 포함\n- 예: `Content-Range: 0-5/50` (전체 50개 중 0-5번째)\n- 페이지네이션 계산에 필요\n\n**count=none** (기본)\n- 전체 개수를 포함하지 않음\n- 응답 속도가 조금 더 빠름",
            "example": "count=exact",
            "schema": {
              "type": "string",
              "enum": ["count=exact", "count=none"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공적으로 게시글을 조회했습니다",
            "headers": {
              "Content-Range": {
                "schema": {
                  "type": "string"
                },
                "description": "조회 범위 및 전체 개수 (예: `0-5/50` = 0번째부터 5번째까지, 전체 50개)\n\n**주의**: `Prefer: count=exact` 헤더 사용 시에만 포함됩니다"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "examples": {
                  "latest": {
                    "summary": "최신순 기본 조회 (6개)",
                    "value": [
                      {
                        "id": 1,
                        "title": "혁신적인 AI 모델이 2024년 기술 판도를 바꿀 예정",
                        "excerpt": "새로운 AI 모델이 출시되면서 기술 산업에 큰 변화가 예상되고 있습니다.",
                        "likes": 1245,
                        "comments": 342,
                        "views": 8932,
                        "created_at": "2026-04-04T05:30:12.752165",
                        "url": "https://reddit.com/r/technology",
                        "categories": {
                          "name": "tech",
                          "icon": "💻"
                        },
                        "sources": {
                          "name": "Reddit",
                          "icon_emoji": "🔴"
                        }
                      },
                      {
                        "id": 2,
                        "title": "클라우드 비용 최적화 전략",
                        "excerpt": "AWS, Azure, GCP의 비용을 30% 절감하는 방법",
                        "likes": 856,
                        "comments": 195,
                        "views": 4520,
                        "created_at": "2026-04-03T14:20:05.123456",
                        "url": "https://dev.to/cloud-tips",
                        "categories": {
                          "name": "tech",
                          "icon": "💻"
                        },
                        "sources": {
                          "name": "Dev.to",
                          "icon_emoji": "⬛"
                        }
                      }
                    ]
                  },
                  "tech_category": {
                    "summary": "기술 카테고리 필터 + 추천순",
                    "value": [
                      {
                        "id": 5,
                        "title": "Python 3.12 새로운 기능",
                        "excerpt": "pattern matching과 type hints 개선사항 분석",
                        "likes": 2145,
                        "comments": 523,
                        "views": 12450,
                        "created_at": "2026-04-02T10:15:30.456789",
                        "url": "https://peps.python.org/pep-0698/",
                        "categories": {
                          "name": "tech",
                          "icon": "💻"
                        },
                        "sources": {
                          "name": "HackerNews",
                          "icon_emoji": "🔶"
                        }
                      }
                    ]
                  },
                  "with_pagination": {
                    "summary": "페이지네이션 (offset=6, 6개)",
                    "description": "Content-Range 헤더: 6-11/50",
                    "value": [
                      {
                        "id": 7,
                        "title": "WebAssembly 성능 최적화",
                        "excerpt": "WASM으로 JavaScript 성능 10배 향상",
                        "likes": 1567,
                        "comments": 289,
                        "views": 6789,
                        "created_at": "2026-04-01T08:45:20.789012",
                        "url": "https://example.com/wasm",
                        "categories": {
                          "name": "tech",
                          "icon": "💻"
                        },
                        "sources": {
                          "name": "Medium",
                          "icon_emoji": "⚫"
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 쿼리 파라미터\n\n**원인**\n- 범위를 벗어난 limit 값\n- 잘못된 order 값\n- 존재하지 않는 카테고리\n\n**예시 응답**\n```json\n{\n  \"code\": \"22P02\",\n  \"message\": \"invalid input syntax\"\n}\n```"
          },
          "500": {
            "description": "서버 오류\n\n**원인**\n- 데이터베이스 연결 오류\n- PostgREST 내부 오류\n\n**해결 방법**\n- 잠시 후 다시 시도\n- 프로젝트 관리자에게 알림"
          }
        }
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "특정 게시글 상세 조회",
        "description": "## 특정 ID의 게시글 전체 데이터를 조회합니다\n\n게시글의 모든 정보와 관련된 카테고리, 출처 데이터를 포함하여 반환합니다.\n\n### 사용 사례\n- 📖 게시글 상세 페이지 로드\n- 🔗 공유 링크에서 특정 게시글 조회\n- 💾 캐싱을 위해 사전 로드\n\n### 주의사항\n- **ID는 필수**: URL에 반드시 포함되어야 합니다\n- **404 응답**: 존재하지 않는 ID는 404 에러 반환\n- **완전한 데이터**: 카테고리, 출처 정보 자동 포함",
        "externalDocs": {
          "description": "게시글 상세 조회 예제",
          "url": "https://github.com/yourrepo/docs/API_GUIDE.md#posts-id"
        },
        "operationId": "getPostById",
        "x-code-samples": [
          {
            "lang": "cURL",
            "summary": "ID 1의 게시글 조회",
            "source": "curl -X GET 'http://localhost:3001/posts?id=eq.1&select=*,categories(*),sources(*)'"
          },
          {
            "lang": "JavaScript",
            "summary": "게시글 상세 조회",
            "source": "async function getPostDetails(postId) {\n  const response = await fetch(\n    `http://localhost:3001/posts?id=eq.${postId}&select=*,categories(*),sources(*)`\n  );\n  const [post] = await response.json();\n  return post;\n}"
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "게시글 ID\n\n- 유효한 게시글 ID 입력\n- 예: `id=eq.1` 또는 PostgREST 쿼리로 `?id=eq.1`\n- 존재하지 않으면 404 반환",
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "게시글 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "examples": {
                  "success": {
                    "summary": "정상 응답",
                    "value": [
                      {
                        "id": 1,
                        "title": "혁신적인 AI 모델이 2024년 기술 판도를 바꿀 예정",
                        "excerpt": "새로운 AI 모델이 출시되면서 기술 산업에 큰 변화가 예상되고 있습니다.",
                        "content": "AI 산업의 최신 트렌드와 기술 발전에 대한 심층 분석...",
                        "likes": 1245,
                        "comments": 342,
                        "views": 8932,
                        "created_at": "2026-04-04T05:30:12.752165",
                        "updated_at": "2026-04-04T05:30:12.752165",
                        "url": "https://reddit.com/r/technology",
                        "author": "tech_enthusiast",
                        "categories": {
                          "name": "tech",
                          "icon": "💻"
                        },
                        "sources": {
                          "name": "Reddit",
                          "icon_emoji": "🔴"
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "게시글을 찾을 수 없습니다\n\n**원인**\n- 잘못된 게시글 ID\n- 삭제된 게시글\n\n**응답 예**\n```json\n[]\n```\n(빈 배열로 반환됨)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "required": [
          "id",
          "title",
          "excerpt",
          "likes",
          "comments",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "게시글 고유 ID"
          },
          "title": {
            "type": "string",
            "description": "게시글 제목",
            "maxLength": 500
          },
          "excerpt": {
            "type": "string",
            "description": "게시글 요약 (미리보기 텍스트)"
          },
          "content": {
            "type": "string",
            "description": "게시글 본문 (전체 내용, 선택적)"
          },
          "likes": {
            "type": "integer",
            "description": "추천 수 / 좋아요 수",
            "minimum": 0
          },
          "comments": {
            "type": "integer",
            "description": "댓글 수",
            "minimum": 0
          },
          "views": {
            "type": "integer",
            "description": "조회 수",
            "minimum": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "게시글 작성 시간 (ISO 8601 형식)"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "게시글 수정 시간 (ISO 8601 형식)"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "게시글 원본 링크 (외부 커뮤니티 URL)"
          },
          "author": {
            "type": "string",
            "description": "게시글 작성자명"
          },
          "categories": {
            "type": "object",
            "description": "카테고리 정보 (관계 데이터)",
            "properties": {
              "name": {
                "type": "string",
                "description": "카테고리명"
              },
              "icon": {
                "type": "string",
                "description": "카테고리 아이콘 (이모지)"
              }
            }
          },
          "sources": {
            "type": "object",
            "description": "출처 정보 (관계 데이터)",
            "properties": {
              "name": {
                "type": "string",
                "description": "출처명"
              },
              "icon_emoji": {
                "type": "string",
                "description": "출처 아이콘 (이모지)"
              }
            }
          }
        }
      },
...
    }
  }
}